+++ project_book/src/main/resources/templates/login/login.html
@@ -6,16 +6,35 @@
           content="width=device-width, user-scalable=no, initial-scale=1">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Login</title>
-    <style>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="format-detection" content="telephone=no">
+    <meta name="apple-mobile-web-app-capable" content="yes">
+    <meta name="author" content="">
+    <meta name="keywords" content="">
+    <meta name="description" content="">
 
+    <link rel="stylesheet" type="text/css" href="/css/normalize.css">
+    <link rel="stylesheet" type="text/css" href="/icomoon/icomoon.css">
+    <link rel="stylesheet" type="text/css" href="/css/vendor.css">
+    <link rel="stylesheet" type="text/css" href="/style.css">
+
+    <!-- script
+    ================================================== -->
+    <script src="/js/modernizr.js"></script>
+    <style>
+        .body{
+            background: rgb(243, 242, 236);
+        }
     </style>
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
           integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 </head>
-<body>
+<body class="body" >
+<div th:insert="user/navbar :: copy"></div>
 
 <!--section-->
-<section class="vh-100" style="background-color: #9A616D;">
+<section class="vh-100" style="background: rgb(243, 242, 236);">
     <div class="container py-5 h-100">
         <div class="row d-flex justify-content-center align-items-center h-100">
             <div class="col col-xl-10">
@@ -42,11 +61,7 @@
                                             You have been logged out successfully.
                                         </strong>
                                     </div>
-                                    <div class="alert alert-success" role="alert" th:if="${param.login}">
-                                        <strong>
-                                            Tài khoản hoặc mật khẩu không đúng
-                                        </strong>
-                                    </div>
+
                                     <div class="form-outline mb-4">
                                         <input type="text" id="form2Example17"
                                                class="form-control form-control-lg"
@@ -71,7 +86,7 @@
 
                                     <a class="small text-muted" href="#!">Forgot password?</a>
                                     <p class="mb-5 pb-lg-2" style="color: #393f81;">Don't have an account?
-                                        <a href="#" style="color: #393f81;">
+                                        <a href="/welcome/register" style="color: #393f81;">
                                             Register here
                                         </a>
                                     </p>
@@ -88,6 +103,8 @@
     </div>
 </section>
 <!--section-->
+<div th:insert="user/footer :: copy"></div>
+
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
         crossorigin="anonymous"></script>
