+++ project_book/src/main/resources/templates/login/register.html
@@ -3,12 +3,30 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="format-detection" content="telephone=no">
+    <meta name="apple-mobile-web-app-capable" content="yes">
+    <meta name="author" content="">
+    <meta name="keywords" content="">
+    <meta name="description" content="">
+
+    <link rel="stylesheet" type="text/css" href="/css/normalize.css">
+    <link rel="stylesheet" type="text/css" href="/icomoon/icomoon.css">
+    <link rel="stylesheet" type="text/css" href="/css/vendor.css">
+    <link rel="stylesheet" type="text/css" href="/style.css">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
+          integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+
+    <!-- script
+    ================================================== -->
+    <script src="/js/modernizr.js"></script>
     <style>
         form {
             width: 400px;
             margin: 0 auto;
             padding: 20px;
-            background-color: #f1f1f1;
+            background-color: white;
             border-radius: 10px;
         }
 
@@ -31,7 +49,6 @@
         }
 
         input[type="submit"] {
-            background-color: #4CAF50;
             color: white;
             padding: 12px 20px;
             border: none;
@@ -44,25 +61,34 @@
             background-color: #45a049;
         }
     </style>
+
 </head>
-<h1 style="text-align: center;background-color: #f1f1f1">THÊM MỚI NGƯỜI DÙNG</h1>
-<body>
+<body style="background: rgb(243, 242, 236)">
+<div th:insert="user/navbar :: copy" ></div>
+<div class="container mb-3">
+    <div class="row">
+        <div class="col-12">
+            <h1 style="text-align: center;background-color: #f1f1f1">Form Register</h1>
+        </div>
+    </div>
+</div>
 <!--        Create: Huynh Duc-->
 <!--        Day: 07/07/2023-->
-<form method="POST" th:action="@{/home/create}" th:object="${add}">
+
+<form method="POST" th:action="@{/welcome/create}" th:object="${add}">
     <div class="user">
         <div>
-            <label for="name">Tên:</label><br>
+            <label for="name">Full name</label><br>
             <input id="name" type="text" th:field="*{name}" placeholder="Dang Tuan Vu"/>
             <span style="color: red" th:if="${#fields.hasErrors('name')}" th:errors="*{name}"></span>
         </div>
         <div>
-            <label for="phoneNumber">Số điện thoại:</label><br>
+            <label for="phoneNumber">Phone:</label><br>
             <input id="phoneNumber" type="text" th:field="*{phone}" placeholder="ex:0123456789"/>
             <span style="color: red" th:if="${#fields.hasErrors('phone')}" th:errors="*{phone}"></span>
         </div>
         <div>
-            <label for="myDate">Ngày Sinh:</label><br>
+            <label for="myDate">Date of birth:</label><br>
             <input type="date" id="myDate" onchange="validateDate()" th:field="*{birthOfDay}">
             <span style="color: red" th:if="${#fields.hasErrors('birthOfDay')}" th:errors="*{birthOfDay}"></span>
         </div>
@@ -74,19 +100,23 @@
         <!--        Create: Huynh Duc-->
         <!--        Day: 07/07/2023-->
         <div>
-            <label>Giới tính:</label>
-            <input type="radio" th:field="*{genderUser}" id="male" value="0"/>Nam
-            <input type="radio" th:field="*{genderUser}" id="female" value="1"/>Nữ
-            <input type="radio" th:field="*{genderUser}" id="other" value=""/>Khác
+            <label>Gender:</label>
+            <input type="radio" th:field="*{genderUser}" id="male" value="0"/>Male
+            <input type="radio" th:field="*{genderUser}" id="female" value="1"/>Female
+            <input type="radio" th:field="*{genderUser}" id="other" value=""/>Other
         </div>
         <div>
-            <label for="pass">Mật khẩu:</label><br>
+            <label for="pass">Password:</label><br>
             <input id="pass" type="text" th:field="*{passUser}"/>
             <span style="color: red" th:if="${#fields.hasErrors('passUser')}" th:errors="*{passUser}"></span>
         </div>
-        <input type="submit"/>
+        <input type="submit" class="mt-4" style="background-color: #DAA55A"/>
     </div>
 </form>
+<div th:insert="user/footer :: copy"></div>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
+        crossorigin="anonymous"></script>
 <script>
     function validateDate() {
         let inputDate = new Date(document.getElementById("myDate").value);
@@ -100,3 +130,4 @@
 </script>
 </body>
 </html>
+
