+++ project_book/src/main/resources/templates/user/navbar.html
@@ -24,7 +24,7 @@
 
 <body>
 
-<div id="header-wrap" th:fragment="copy" th:object="${cart}" >
+<div id="header-wrap" th:fragment="copy" th:object="${cart}">
 
     <div class="top-content">
         <div class="container">
@@ -47,10 +47,18 @@
                         </ul>
                     </div><!--social-links-->
                 </div>
+
                 <div class="col-md-6">
-                    <div class="right-element">
-                        <a href="#" class="user-account for-buy"><i class="icon icon-user"></i><span>Account</span></a>
-                        <a href="/cart/show-cart" class="cart for-buy"><i class="icon icon-clipboard"></i><span th:text="'Cart: '+${cart.getItems().size()}"></span></a>
+
+                    <!--                Create: Huynh Duc-->
+                    <!--                Day: 06/07/2023-->
+                    <div class="right-element" th:if="${#request.userPrincipal!=null}" th:object="${user}">
+                        <i class="icon icon-user"></i>
+                        <span th:text="${user.name}" >
+                        </span>
+                        <a href="/welcome/logout" class="user-account for-buy"><span>Logout</span></a>
+                        <a href="/cart/show-cart" class="cart for-buy"><i class="icon icon-clipboard"></i><span
+                                th:text="'Cart: '+${cart.getItems().size()}"></span></a>
 
                         <div class="action-menu">
 
@@ -64,7 +72,29 @@
                             </div>
                         </div>
 
-                    </div><!--top-right-->
+                    </div>
+                    <!--                Create: Huynh Duc-->
+                    <!--                Day: 06/07/2023-->
+                    <div class="right-element" th:unless="${#request.userPrincipal!=null}">
+
+                        <a href="/welcome/login" class="user-account for-buy"><i class="icon icon-user"></i><span>Account</span></a>
+                        <a href="/cart/show-cart" class="cart for-buy"><i class="icon icon-clipboard"></i><span
+                                th:text="'Cart: '+${cart.getItems().size()}"></span></a>
+
+                        <div class="action-menu">
+
+                            <div class="search-bar">
+                                <a href="#" class="search-button search-toggle" data-selector="#header-wrap">
+                                    <i class="icon icon-search"></i>
+                                </a>
+                                <form role="search" method="get" class="search-box">
+                                    <input class="search-field text search-input" placeholder="Search" type="search">
+                                </form>
+                            </div>
+                        </div>
+
+                    </div>
+                    <!--top-right-->
                 </div>
 
             </div>
@@ -87,10 +117,13 @@
                         <div class="main-menu stellarnav">
                             <ul class="menu-list">
                                 <li class="menu-item active"><a th:href="@{/welcome}" data-effect="Home">Home</a></li>
-                                <li class="menu-item"><a href="#about" class="nav-link" data-effect="About">About</a></li>
+                                <li class="menu-item"><a href="#about" class="nav-link" data-effect="About">About</a>
+                                </li>
                                 <li class="menu-item has-sub">
-                                <li class="menu-item"><a th:href="@{/welcome/view-all}" class="nav-link" data-effect="Shop">Shop</a></li>
-                                <li class="menu-item"><a href="#latest-blog" class="nav-link" data-effect="Articles">Articles</a></li>
+                                <li class="menu-item"><a th:href="@{/welcome/view-all}" class="nav-link"
+                                                         data-effect="Shop">Shop</a></li>
+                                <li class="menu-item"><a href="#latest-blog" class="nav-link" data-effect="Articles">Articles</a>
+                                </li>
                             </ul>
 
                             <div class="hamburger">
